Moves WeightedStatsRSocketProxy to toplevel package protected class.
Moves wrap() method to WeightedStats default method.
