+++ rsocket-core/src/main/java/io/rsocket/loadbalance/WeightedStats.java
@@ -1,7 +1,9 @@
 package io.rsocket.loadbalance;
 
+import io.rsocket.RSocket;
+
 /**
- * Representation of stats used by the {@link WeightedLoadbalanceStrategy}
+ * Representation of stats used by the {@link WeightedLoadbalanceStrategy}.
  *
  * @since 1.1
  */
@@ -16,4 +18,15 @@ public interface WeightedStats {
   double predictedLatency();
 
   double weightedAvailability();
+
+  /**
+   * Wraps an RSocket with a proxy that implements WeightedStats.
+   *
+   * @param rsocket the RSocket to proxy.
+   * @return the wrapped RSocket.
+   * @since 1.1.1
+   */
+  default RSocket wrap(RSocket rsocket) {
+    return new WeightedStatsRSocketProxy(rsocket, this);
+  }
 }
